# Conflicts:
#	shop/frontend/src/app/components/Component-function/site-header/site-header.component.spec.ts
#	shop/frontend/src/app/components/Component-login/forgot-password/forgot-password.component.css
#	shop/frontend/src/app/components/Component-login/forgot-password/forgot-password.component.html

+++ shop/backend/src/main/java/com/g5/techdevices/techstore/repositories/cart/ProductVariantRepository.java
@@ -3,5 +3,5 @@ package com.g5.techdevices.techstore.repositories.cart;
 import com.g5.techdevices.techstore.entity.products.ProductVariant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductVariantRepository extends JpaRepository<ProductVariant, Integer> {
+public interface ProductVariantRepository extends JpaRepository<ProductVariant, Long> {
 }
